AIチャット関連の型定義を整理し新規追加
