+++ src/features/ai-chat/types.ts
@@ -1,4 +1,20 @@
-import type { AiChatToolTrace, OllamaErrorDetails } from '@/types/background'
+import type {
+  AiChatAttachment,
+  AiChartSpec,
+  AiChatToolTrace,
+  OllamaErrorDetails,
+} from '@/types/ai-chat-protocol'
+
+export type {
+  AiChatAttachment,
+  AiChartAxisFormat,
+  AiChartDatum,
+  AiChartSeries,
+  AiChartSpec,
+  AiChartType,
+  AiChatToolTrace,
+  OllamaErrorDetails,
+} from '@/types/ai-chat-protocol'
 
 export interface AiSavedUrlRecord {
   id: string
@@ -41,13 +57,6 @@ export interface AiSavedUrlPage<T> {
   sortDirection: AiSavedUrlSortDirection
 }
 
-export interface AiChatAttachment {
-  filename: string
-  mediaType: string
-  kind: 'text' | 'image'
-  content: string
-}
-
 export interface AiChatConversationMessage {
   attachments?: AiChatAttachment[]
   charts?: AiChartSpec[]
@@ -80,30 +89,19 @@ export interface InterestEvidenceEntry {
   count: number
 }
 
-export type AiChartType = 'area' | 'bar' | 'line' | 'pie' | 'radar'
-
-export type AiChartAxisFormat = 'count' | 'date' | 'label' | 'percent'
-
-export interface AiChartSeries {
-  colorToken: string
-  dataKey: string
-  label: string
+export interface AiChatConversation {
+  createdAt: number
+  id: string
+  messages: AiChatConversationMessage[]
+  title: string
+  updatedAt: number
 }
 
-export type AiChartDatum = Record<string, number | string | null>
-
-export interface AiChartSpec {
-  type: AiChartType
+export interface AiChatHistoryItem {
+  id: string
+  isActive: boolean
+  preview: string
   title: string
-  data: AiChartDatum[]
-  series: AiChartSeries[]
-  categoryKey?: string
-  description?: string
-  emptyMessage?: string
-  showLegend?: boolean
-  stacked?: boolean
-  valueFormat?: AiChartAxisFormat
-  xKey?: string
 }
 
 export interface InterestInferenceResult {
